+++ packages/react/src/hooks/index.ts
@@ -30,7 +30,10 @@ export { type UseRemoteParticipantOptions, useRemoteParticipant } from './useRem
 export { type UseRemoteParticipantsOptions, useRemoteParticipants } from './useRemoteParticipants';
 export { type UseRoomInfoOptions, useRoomInfo } from './useRoomInfo';
 export { useSortedParticipants } from './useSortedParticipants';
-export { useSpeakingParticipants } from './useSpeakingParticipants';
+export {
+  useSpeakingParticipants,
+  type UseSpeakingParticipantsOptions,
+} from './useSpeakingParticipants';
 export { type UseStartAudioProps, useStartAudio } from './useStartAudio';
 export { type UseStartVideoProps, useStartVideo } from './useStartVideo';
 export { type UseSwipeOptions, useSwipe } from './useSwipe';
@@ -57,3 +60,16 @@ export * from './useIsRecording';
 export { useMediaQuery } from './internal'; // Added this line by akash
 export { useWarnAboutMissingStyles } from './useWarnAboutMissingStyles'; // Added this line by akash
 export { useObservableState } from './internal/useObservableState'; // Added this line by akash
+export * from './useTextStream';
+export * from './useTranscriptions';
+export * from './useSequentialRoomConnectDisconnect';
+export * from './useSession';
+export {
+  type AgentState,
+  AgentEvent,
+  type AgentCallbacks,
+  type UseAgentReturn,
+  useAgent,
+} from './useAgent';
+export * from './useEvents';
+export * from './useSessionMessages';